Write meter data summary table.
Add comments describing the meterddata table.
Create the table if it done not exist in the target database.
Import specific functions/classes from ekmmeters instead of import *.
Add comments describing the database trigger function that creates rows in meterdata.
Function "makeMeterDataMsg" returns a dictionary of results instead of a string.